回転・拡縮に対応するべく SpringBone joint の TRS を質点の位置に解決するように整理する
vrm0 > vrm0.fast > vrm1.fast の順に作業する。３つとも処理内容は同じ
