+++ Assets/VRM/Runtime/SpringBone/Logic.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1b3d9c467d3da3e4aa800f2a9229768c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
