+++ Assets/VRM/Runtime/SpringBone/Logic/SceneInfo.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace VRM.SpringBone
+{
+    struct SceneInfo
+    {
+        public IReadOnlyList<Transform> RootBones;
+        public Transform m_center;
+        public VRMSpringBoneColliderGroup[] ColliderGroups;
+    }
+}
