+++ Assets/VRM/Runtime/SpringBone/Logic/SceneInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f9f3d46865885b4a8d559a5d962db5a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
