+++ Assets/VRM/Runtime/SpringBone/Logic/SphereCollider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 102ef8d07b335ee42b6a198b5657ef80
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
