+++ Assets/VRM/Runtime/SpringBone/Logic/SpringBoneJoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2602939771f9eb2428b140e10fa899f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
