+++ Assets/VRM/Runtime/SpringBone/Logic/SpringBoneSettings.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+namespace VRM.SpringBone
+{
+    struct SpringBoneSettings
+    {
+        public float m_stiffnessForce;
+        public Vector3 m_gravityDir;
+        public float m_gravityPower;
+        public float m_hitRadius;
+        public float m_dragForce;
+    }
+}
