+++ Assets/VRM/Runtime/SpringBone/Logic/SpringBoneSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c979a822d9cd3754db84ce10c42f3b56
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
