+++ Assets/VRM/Runtime/SpringBone/Logic/SpringBoneSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9cc72abc9e20895419ca606484cc7dc0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
